Implementing Rewards to the GA, adjusted some of the weights of penalties and rewards.
now taking random individuals to fill free spaces in select parents fucntion instead of random generated ones which would overwhelm the others.

movement based rewards are based on how many legs are moving and averaging all the legs together.

issue now, is the first few frames all legs are moving but they all come to a stop, we are reaching local minimums of reaching the maximum values then not being able to move out from it due to maximums and deltas or not wanting to intersect etc